librdkafka bump and only use JSON statistics for now to maintain ABI compat
