+++ src/client.rs
@@ -31,8 +31,8 @@ use crate::groups::GroupList;
 use crate::log::{debug, error, info, trace, warn, LogRecord};
 use crate::metadata::Metadata;
 use crate::mocking::MockCluster;
-use crate::statistics::StatsView;
 use crate::util::{self, ErrBuf, KafkaDrop, NativePtr, Timeout};
+use crate::Statistics;
 
 /// Client-level context.
 ///
@@ -85,29 +85,28 @@ pub trait ClientContext: Send + Sync {
         }
     }
 
-    #[doc(hidden)]
-    fn stats_view(&self, stats: StatsView<'_>) {
-        info!(
-            "Client stats: name={}, client_id={}, type={}, msg_cnt={}, msg_size={}",
-            stats.name(),
-            stats.client_id(),
-            stats.client_type().as_str(),
-            stats.msg_cnt(),
-            stats.msg_size()
-        );
+    /// Receives the decoded statistics of the librdkafka client. To enable, the
+    /// `statistics.interval.ms` configuration parameter must be specified.
+    ///
+    /// The default implementation logs the statistics at the `info` log level.
+    fn stats(&self, statistics: Statistics) {
+        info!("Client stats: {:?}", statistics);
     }
+
+    // Keep the typed statistics view API unhooked for now. The active stats
+    // delivery path remains the historical JSON callback via `stats_raw`.
 
     /// Receives the JSON-encoded statistics of the librdkafka client. To
     /// enable, the `statistics.interval.ms` configuration parameter must be
     /// specified.
     ///
-    /// The default implementation logs the raw statistics JSON at the `info`
-    /// log level.
+    /// The default implementation calls [`ClientContext::stats`] with the
+    /// decoded statistics, logging an error if the decoding fails.
     fn stats_raw(&self, statistics: &[u8]) {
-        info!(
-            "Client stats raw JSON: {}",
-            String::from_utf8_lossy(statistics)
-        );
+        match serde_json::from_slice(statistics) {
+            Ok(stats) => self.stats(stats),
+            Err(e) => error!("Could not parse statistics JSON: {}", e),
+        }
     }
 
     /// Receives global errors from the librdkafka client.
@@ -373,14 +372,8 @@ impl<C: ClientContext> Client<C> {
     }
 
     fn handle_stats_event(&self, event: *mut RDKafkaEvent) {
-        // Try typed stats first (more efficient, no JSON parsing)
-        let stats_ptr = unsafe { rdsys::rd_kafka_event_stats_typed(event) };
-        if !stats_ptr.is_null() {
-            let view = unsafe { StatsView::new(&*stats_ptr) };
-            self.context().stats_view(view);
-            return;
-        }
-        // Fall back to JSON parsing if typed stats not available
+        // Keep stats delivery on the historical JSON callback path so binaries
+        // do not gain a hard dependency on newer librdkafka symbols.
         let json = unsafe { CStr::from_ptr(rdsys::rd_kafka_event_stats(event)) };
         self.context().stats_raw(json.to_bytes());
     }
